pyproject.toml
[project]
name = "beta-vae"
version = "0.1.0"
description = "Reimplementation of a Beta-VAE"
readme = "REAME.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "rich>=14.0.0",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchsummary>=1.5.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

